+++ src/main/java/edu/hw8/task1/Client.java
@@ -3,27 +3,34 @@ package edu.hw8.task1;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
 import java.net.Socket;
 
 public class Client {
     private static final String HOST = "localhost";
-    private static final int PORT = 8080;
+    private static final int PORT = 1234;
 
     public void startConnection() {
         try (
             Socket socket = new Socket(HOST, PORT);
-            PrintWriter writer = new PrintWriter(socket.getOutputStream(), true);
-            BufferedReader reader = new BufferedReader(new InputStreamReader(socket.getInputStream()));
-            BufferedReader stdIn = new BufferedReader(new InputStreamReader(System.in))
+            BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
+            PrintWriter writer = new PrintWriter(new OutputStreamWriter(socket.getOutputStream()));
+            BufferedReader serverReader = new BufferedReader(new InputStreamReader(socket.getInputStream()))
         ) {
-            String line;
-            while ((line = stdIn.readLine()) != null) {
-                writer.println(line);
-                System.out.println("Server: " + reader.readLine());
+            System.out.println("Введите ключевое слово:");
+            String keyword;
+            while (!"close".equals(keyword = reader.readLine())) {
+                writer.println(keyword);
+                writer.flush();
+                System.out.println("Server: " + serverReader.readLine());
             }
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    public static void main(String[] args) {
+        new Client().startConnection();
     }
 }
