+++ src/main/java/edu/hw8/task1/Server.java
@@ -7,7 +7,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 public class Server extends Thread {
-    private static final int PORT = 8080;
+    private static final int PORT = 1234;
     private static final int MAX_CONNECTIONS = 5;
 
     public void startServer() {
@@ -22,5 +22,9 @@ public class Server extends Thread {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    public static void main(String[] args) {
+        new Server().startServer();
     }
 }
