+++ src/main/java/edu/hw8/task1/Task1.java
@@ -23,10 +23,5 @@ package edu.hw8.task1;
  */
 
 public class Task1 {
-    public static void main(String[] args) {
-        Server server = new Server();
-        server.startServer();
-
-    }
 
 }
